server/requirements.txt
mysql-connector-python
roslibpy
